pyproject.toml
[tool.poetry]
name = "eataway"
version = "0.1.0"
description = "A filter which eats the content of an image away."
authors = ["moodule <moodule@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.2"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
